Update view specs to match filters production code
